Refine web app analytics increments
